Fix ITextBuffer leak after RenameTracking commit
Fixes #830

When a rename tracking session is committed, we create 4
ITextUndoPrimitives to facilitate the two-phase undo system that can
clear or restore an earlier TrackingSession. These UndoPrimitives can
live beyond the lifetime of the buffer in which the TrackingSession
occurred (when they are grouped into a global undo transaction), so it's
important that they do not contain any strong references to the
ITextBuffer.

The UndoPrimitive class previously held a strong reference to the
StateMachine, which directly referenced the ITextBuffer. Instead of
directly referencing the StateMachine, the UndoPrimitive class now holds
a WeakReference to the ITextBuffer, from which we can retrieve the
StateMachine (via its property bag). If the ITextBuffer has gone away
but the UndoPrimitive still exists as part of a global undo, then
undo/redo will not be able to retrieve the StateMachine and will
therefore have no effect (as expected, because the only purpose of the
custom UndoPrimitives is to provide custom experiences in the
ITextBuffer that's already been GC'd).

The UndoPrimitive class also previously referenced the TrackingSession
that it needs to restore on undo, which then contains an ITrackingSpan
that references the ITextBuffer. Instead, the StateMachine itself now
keeps a list of all of the committed TrackingSessions, and the
UndoPrimitive simply stores the index of it's target TrackingSpan into
that list. Because the StateMachine is only held alive in the
ITextBuffer's property bag, the StateMachine itself and all of the
previously committed TrackingSessions will be collected when the
ITextBuffer goes away.