Object Creation Preselection special case
There are cases in broken code scenarios where the new keyword in
objectcreationexpression appears to be a part of a subsequent
assignment.

This was originally addressed in #2537. This commit addresses #2836
which is a special case of the same class of problem.

E.g:
```C#
bool b;
Program p = new $$
b = false;
```

Per the parser, the new token here belongs to the second assignment and
hence type inferrer recommends `bool`.  However, that is not helpful.
The IDE detects this and correctly infers the type `Program` which is
what the user expects.