Move solutions into the root directory
It's confusing for new users to have the Roslyn solutions nested
inside the src/ directory and there's really no good reason those where
there in the first place. This moves them to the root directory, except
for the Toolset.sln solution, which is only for building Roslyn in CI
scenarios and has been moved to the build/ directory.