Add core CLR projects for csc and vbc.
Adds two new projects, CscCore and VbcCore, which target CoreCLR and
produce CoreCLR-compatible executables in the core-clr subdirectory of
the build output. At the moment they are not included in any solution
because a bug in the DNX build tools that causes VS 2015 to deadlock
whenever the project files are loaded.