Add two new dev solutions to the root
Right now we have CrossPlatform, Roslyn, and RoslynLight, all of
which are heavyweight except for CrossPlatform, but xplat isn't
a fixed set of projects, just whatever we can build xplat. This change
adds two new solutions for developers to choose, depending on what
they're working on. The Workspaces solution contains everything
workspaces and down, while the compilers solution just contains the C#
and VB compilers.