Added a document that lists some CodeGen differences.
There is a number of known changes in the binaries produced for the same source code when compared to equivalent output of the old compiler. Some changes are improvements, some are bug fixes and some are indirect result of differences in implementation and emit strategies.

Most of the actual form and shape of emitted code is not a contract and compiler is only responsible for efficiently conveying the semantics of the original source. However there is enough interest in intentional changes here and it seems to be useful to have one place where some of the more significant changes are listed.