Make TokenOfStringCached  independent of length
For lengths that are not between 2 and 16 (inclusive) it immediately
returned its second argument.  For all other cases, it inspected the
string.  When the second argument was anything other than
SyntaxKind.Identifier (in particular, SyntaxKind.XmlNameToken), the
two values would not agree and so the output depended on the length
of the first (string) argument.

For example, passing ("a", SyntaxKind.XmlNameToken) used to return
SyntaxKind.XmlNameToken, but passing ("aa", SyntaxKind.XmlNameToken)
used to return SyntaxKind.IdentifierToken.  Now they both return
SyntaxKind.XmlNameToken.

The only impact seems to be on parsing error recovery.  If this
turns out to make the behavior less desirable, then then we should
change the behavior so that *both* return SyntaxKind.IdentifierToken.