Introduce CompletionItemRules
This change touches a lot of files but most of the changes are
mechanical. Essentially, this introduces a CompletionItemRules type as a
means to customize the behavior of a CompletionItem. This will
ultimately be the home of several methods that currently reside on
CompletionListProvider (such as IsCommitCharacter and
IsFilterCharacter), but only the SendEnterThroughToEditor method has
been moved in this change. Other methods will be moved in future
changes.