Separate UI-thread-only code
Wrap all members that can only be accessed on the UI thread in an inner
class.  (Unfortunately, a struct was too complicated because capturing
```this``` resulted in copying that hid side effects.)

Remove all asserts about running on the UI thread.  Now that all accesses
are funnelled through a single field (and there are only a dozen!), it's
easy to verify by inspection that it is only accessed on the UI thread.

Add unit tests for both the origin bug and for further exploring the
threading behavior of the type (added a bunch of marshalling in response
to exceptions encountered while testing).

Eliminated some dead code.