Test for structs in using in async method
This issue was reported through the following stackoverflow post:

>
http://stackoverflow.com/questions/29230626/why-is-enumerator-movenext-not-working-as-i-expect-it-when-used-with-using-and-a

The Roslyn compiler fixed a bug present in the native compiler around
the use of structs inside a using inside an async method.  The native
compiler was generating some of the method calls on a copy of the struct
which caused  the mutation to be ignored.   Roslyn correctly generates
them all on the lifted value.

Looking around I couldn't find a specific regression test for this
functionality so adding one.