Fix VB member body span used to determine whether speculative binding should be used
To determine whether or not speculative binding should be used, GetSemanticModelForSpanAsync calls into ISyntaxFactsService.GetMemberBodySpanForSpeculativeBinding(). In VB, the span returned by this method started from the end of the block statement to the start of the end block statement. However, for VB, the span should start after the block statement's terminator. Otherwise, a member body speculative semantic model will likey be returned when the requested span is within the trivia after the block statement but before the statement terminator. E.g.

Sub M() |

End Sub