Fix VB TypeInferrer after ArgumentList
Fixes #3518

When trying to figure out which Argument of an ArgumentList was being
inferred, we found the index and then returned "(index + 1) \ 2" without
checking if the index was out of bounds. When the index is out of bounds
its value is -1, so the above expression results in an index of 0, and
the rest of the system continues as though we're really trying to infer
the type of the first Argument (but really, the token we've passed is
the close paren for the entire invocation).