Fix NRE in command line parsing
The command line parser was inconsistent in how it set the
sourceFilesSpecified boolean flag.  In one place it tracked whether the
source files collection was non empty and in another it tracked whether
or not we had attempted to parse an argument as a source file.  This
disconnect lead to a NRE when there was an invalid source file and the
command line tried to calculate the output name for a library project.
Changed the compiler to consistently treat it as tracking whether or not
a valid source file was specified.