Lifted IntPtr Conversions
The compiler wasn't properly handling the case of a conversion between
Nullable<int> and Nullable<IntPtr>.  The compiler specially handles
IntPtr conversions but wasn't properly lowering them in the
LocalRewriter.

I verified the output IL and expression trees are identical to the
native compiler for these cases.

This closes DevDiv 1210529