Fix VB formatting to stop introducing a space before the ? in a conditional access expression
The code to handle spaces before the ? in a conditional access expression was very specific, only removing the space if the previous token was of a specific kind. However, there were several cases missed (and one bug after MyClass). Now we take a more general approach that is similar to how we treat member-access expressions: If we see a conditional access expression, we assume that there should never be a space before the ?. Any cases where a space is desired would be exceptional.