Be more conservative about formatting braces across multiple lines when the tree has missing curlies.
This is to ensure that if you have

```cs
class C
{
    int Foo { get { return
}
```

And you type ```;``` after ```return``` that we don't place the get-accessor on multiple lines.
In this case the parser thinks the last curly brace belongs to the get-accessor (and thus we
used to naively think the get-accessor was on multiple lines).

Note: we customize our behavior here depending on which token you're typing.  If you're actually
typeing the close curly, then we assume the curly belongs to the nearest construct (and thus in
the above example the 'get-accessor' *would* be on multiple lines).  This is because it's unlikely
that you would have left the accessor in this unfinished state to go work on some other construct.

However, if you're not actually typing the close curly (and there are misisng close curlies), then
we'll conservatively assume that the close curly might not belong to us when determining if we
were written on a single line.

Note: most of the change here is simply due to having to pass along the information about what
token was being edited.