Get document moniker from IVsProject instead of ROT
Switching to IVsProject.GetMkDocument to get document moniker instead of
iterating through all the entries in the running object table. Also
cleaning up a bit of code while I'm here - It looks like we don't need
to determine Shared Hierarchy & Shared ItemId here anymore. It was there
for error list and it looks like with PR #2399 error list switched to a
different implementation.