Enable selection of Mef components based on TextViewRole...
We need a way to provide alternate completion in the Interactive Window.
Previously, we weren't taking TextViewRole into consideration when
selecting completion providers.  This change allows a service to opt in to
considering TextViewRole by passing OrderableLanguageAndRoleMetadata to
SelectMatchingExtensionValues.

This change also includes some cleanup of test helpers and removal of extra
TextViewRole attributes that weren't doing anything meaningful.