Check all owning nodes instead of just the first
Handle cases where a block syntax could be parented by an Arrow clause
syntax.  In such cases the first owner would be a statement syntax,
where introducing a line ender would produce a syntax error. If we
considered the larger case where the statement is parented by an
expression bodied member, we wouldn't be introducing any errors
actually. So, consider the set of owning nodes from the most immediate
parent and outward. If introducing a line ender at any such level
succeeds, accept that. We're still conservative enough to not introduce
any errors in correct code.