Handle ConditionalAccessExpression cases properly.
SyntaxFacts.GetExpressionOfMemberAccessExpression was looking for
patterns that could be ConditionalAccessExpression. It was looking only
two levels up from the current node. We hit a case where the
ConditionalAccessExpression was four levels higher. So, in general,
instead of dealing with various cases, we should use the
GetParentCondtionalAccessExpression helper that walks up from a given
node and finds a ConditionalAccessExpression if one exists in a parent
hierarchy. This change is C# only. The VB implementation of this method
correctly uses GetCorrespondingConditionalAccessExpression in this case.